requirements.txt
asgiref==3.7.2
astroid==3.1.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
click==8.1.7
dill==0.3.8
Django==5.0.3
django-debug-toolbar==4.3.0
django-filter==24.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
docutils==0.21.2
drf-spectacular==0.27.2
flake8==7.0.0
flake8-docstrings==1.7.0
gunicorn==22.0.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sentry-sdk==2.3.1
snowballstemmer==2.2.0
sqlparse==0.4.4
tomlkit==0.12.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
